// src/models/fixed-pool-model.ts
export interface FixedPoolModel {
  _id?: string
  ownerLink?: string
  description?: string
  endDate?: string
  startDate?: string
  accessType?: string
  amount?: string
  ratio?: number
  chainId?: number
  chain?: string
  totalSupply?: number
  totalRaise?: number
  tokenAddress?: string
  tokenName?: string
  maxAllocation?: number
  minAllocation?: number
  status?: string
  type?: 'v1' | 'v2' | 'v3' | 'v4' | 'v5'
  name?: string
  slugName?: string
  createdAt?: Date
  updatedAt?: Date
  __v?: number
  file?: any
  address?: string
  tradeToken?: string
  tradeTokenAddress?: string
  id?: string
  logoUrl?: string
  coverImageUrl?: string
  categoryPool?: 'ido' | 'invest'
  index?: number
  vestingAddress?: string
  vestingNetwork?: string
  vestingChainId?: number
  vestingTokenDecimal?: number
  data?: {
    baseParticipants?: number
    participants?: number
    tokensAllocated?: string
    minAllocationUsd?: number
    maxAllocationUsd?: number
    totalRaiseUsd?: number
    shortDescription?: string
    medium?: string
    telegram?: string
    telegramChat?: string
    twitter?: string
    facebook?: string
    instagram?: string
    youtube?: string
    discord?: string
    reddit?: string
    github?: string
    web?: string
    whitelistUrl?: string
    forceProgressPercent?: number
    baseProgressPercent?: number
    claimType?: 'airdrop'
    startDate: string
    isTBARedeem?: boolean
    solanaRequired?: boolean
    tax?: number
    showRefund?: boolean
    hideParticipants?: boolean
    hideContractAddress?: boolean
    forceRefund?: boolean
    isExclusive?: boolean
  }
}